Add support for compression.
This only adds supports for those clients following RFC 7692 (i.e. those
clients passing and supporting the "permessage-deflate" extension):

    https://tools.ietf.org/html/rfc7692

Those webkit clients using the "x-webkit-deflate-frame" (e.g. safari)
extension are not supported:

    https://tools.ietf.org/id/draft-tyoshino-hybi-websocket-perframe-deflate-05.txt

One global compressor and decompressor is used to save on memory
overhead.  To accomplish this, the server always limits the compression
context to be on a per-frame basis, so context is not needed to be
maintained for each individual client.